Used longlong instead of int to store larger nums

+++ 0x0C-more_malloc_free/101-mul.c
@@ -19,7 +19,7 @@ int main(int argc, char *argv[])
 
 	else
 	{
-		printf("%d\n", (atoi(argv[1]) * atoi(argv[2])));
+		printf("%lld\n", (atoll(argv[1]) * atoll(argv[2])));
 		return (0);
 	}
 }
